src/BookInfoRetrievalService/requirements.txt
flask==3.0.0
flask-consulate==0.2.0
SQLAlchemy==2.0.23
python-dotenv==1.0.0